refactor: 827 - fix imports
