+++ src/widgets/header/ui/burger/burger.module.scss
