+++ src/widgets/header/ui/dropdown/dropdown-wrapper.module.scss
